+++ pg_documentdb/include/infrastructure/job_management.h
@@ -16,6 +16,7 @@
 
 extern bool EnableBackgroundWorker;
 extern bool EnableBackgroundWorkerJobs;
+extern bool EnableBackgroundWorkerInitJobs;
 extern bool EnableBgWorkerMetricsEmission;
 extern bool IndexBuildsScheduledOnBgWorker;